refactor: rename selection-to-markdown to extract-selection-html
The content script no longer performs Markdown conversion (that moved to the offscreen document / Event Page in #259); it only extracts the selection as an HTML fragment. Rename the file to match its sole export, extractSelectionHtml, and update all importers.
